+++ src/backend/core/tests/documents/test_api_documents_descendants.py
@@ -1,5 +1,5 @@
 """
-Tests for Documents API endpoint in impress's core app: children list
+Tests for Documents API endpoint in impress's core app: descendants
 """
 
 import random
@@ -14,17 +14,19 @@ from core import factories
 pytestmark = pytest.mark.django_db
 
 
-def test_api_documents_children_list_anonymous_public_standalone():
-    """Anonymous users should be allowed to retrieve the children of a public document."""
+def test_api_documents_descendants_list_anonymous_public_standalone():
+    """Anonymous users should be allowed to retrieve the descendants of a public document."""
     document = factories.DocumentFactory(link_reach="public")
     child1, child2 = factories.DocumentFactory.create_batch(2, parent=document)
+    grand_child = factories.DocumentFactory(parent=child1)
+
     factories.UserDocumentAccessFactory(document=child1)
 
-    response = APIClient().get(f"/api/v1.0/documents/{document.id!s}/children/")
+    response = APIClient().get(f"/api/v1.0/documents/{document.id!s}/descendants/")
 
     assert response.status_code == 200
     assert response.json() == {
-        "count": 2,
+        "count": 3,
         "next": None,
         "previous": None,
         "results": [
@@ -38,12 +40,30 @@ def test_api_documents_children_list_anonymous_public_standalone():
                 "is_favorite": False,
                 "link_reach": child1.link_reach,
                 "link_role": child1.link_role,
-                "numchild": 0,
+                "numchild": 1,
                 "nb_accesses_ancestors": 1,
                 "nb_accesses_direct": 1,
                 "path": child1.path,
                 "title": child1.title,
                 "updated_at": child1.updated_at.isoformat().replace("+00:00", "Z"),
+                "user_roles": [],
+            },
+            {
+                "abilities": grand_child.get_abilities(AnonymousUser()),
+                "created_at": grand_child.created_at.isoformat().replace("+00:00", "Z"),
+                "creator": str(grand_child.creator.id),
+                "depth": 3,
+                "excerpt": grand_child.excerpt,
+                "id": str(grand_child.id),
+                "is_favorite": False,
+                "link_reach": grand_child.link_reach,
+                "link_role": grand_child.link_role,
+                "numchild": 0,
+                "nb_accesses_ancestors": 1,
+                "nb_accesses_direct": 0,
+                "path": grand_child.path,
+                "title": grand_child.title,
+                "updated_at": grand_child.updated_at.isoformat().replace("+00:00", "Z"),
                 "user_roles": [],
             },
             {
@@ -68,9 +88,9 @@ def test_api_documents_children_list_anonymous_public_standalone():
     }
 
 
-def test_api_documents_children_list_anonymous_public_parent():
+def test_api_documents_descendants_list_anonymous_public_parent():
     """
-    Anonymous users should be allowed to retrieve the children of a document who
+    Anonymous users should be allowed to retrieve the descendants of a document who
     has a public ancestor.
     """
     grand_parent = factories.DocumentFactory(link_reach="public")
@@ -81,13 +101,15 @@ def test_api_documents_children_list_anonymous_public_parent():
         link_reach=random.choice(["authenticated", "restricted"]), parent=parent
     )
     child1, child2 = factories.DocumentFactory.create_batch(2, parent=document)
+    grand_child = factories.DocumentFactory(parent=child1)
+
     factories.UserDocumentAccessFactory(document=child1)
 
-    response = APIClient().get(f"/api/v1.0/documents/{document.id!s}/children/")
+    response = APIClient().get(f"/api/v1.0/documents/{document.id!s}/descendants/")
 
     assert response.status_code == 200
     assert response.json() == {
-        "count": 2,
+        "count": 3,
         "next": None,
         "previous": None,
         "results": [
@@ -101,12 +123,30 @@ def test_api_documents_children_list_anonymous_public_parent():
                 "is_favorite": False,
                 "link_reach": child1.link_reach,
                 "link_role": child1.link_role,
-                "numchild": 0,
+                "numchild": 1,
                 "nb_accesses_ancestors": 1,
                 "nb_accesses_direct": 1,
                 "path": child1.path,
                 "title": child1.title,
                 "updated_at": child1.updated_at.isoformat().replace("+00:00", "Z"),
+                "user_roles": [],
+            },
+            {
+                "abilities": grand_child.get_abilities(AnonymousUser()),
+                "created_at": grand_child.created_at.isoformat().replace("+00:00", "Z"),
+                "creator": str(grand_child.creator.id),
+                "depth": 5,
+                "excerpt": grand_child.excerpt,
+                "id": str(grand_child.id),
+                "is_favorite": False,
+                "link_reach": grand_child.link_reach,
+                "link_role": grand_child.link_role,
+                "numchild": 0,
+                "nb_accesses_ancestors": 1,
+                "nb_accesses_direct": 0,
+                "path": grand_child.path,
+                "title": grand_child.title,
+                "updated_at": grand_child.updated_at.isoformat().replace("+00:00", "Z"),
                 "user_roles": [],
             },
             {
@@ -132,14 +172,15 @@ def test_api_documents_children_list_anonymous_public_parent():
 
 
 @pytest.mark.parametrize("reach", ["restricted", "authenticated"])
-def test_api_documents_children_list_anonymous_restricted_or_authenticated(reach):
+def test_api_documents_descendants_list_anonymous_restricted_or_authenticated(reach):
     """
-    Anonymous users should not be able to retrieve children of a document that is not public.
+    Anonymous users should not be able to retrieve descendants of a document that is not public.
     """
     document = factories.DocumentFactory(link_reach=reach)
-    factories.DocumentFactory.create_batch(2, parent=document)
+    child = factories.DocumentFactory(parent=document)
+    _grand_child = factories.DocumentFactory(parent=child)
 
-    response = APIClient().get(f"/api/v1.0/documents/{document.id!s}/children/")
+    response = APIClient().get(f"/api/v1.0/documents/{document.id!s}/descendants/")
 
     assert response.status_code == 401
     assert response.json() == {
@@ -148,11 +189,11 @@ def test_api_documents_children_list_anonymous_restricted_or_authenticated(reach
 
 
 @pytest.mark.parametrize("reach", ["public", "authenticated"])
-def test_api_documents_children_list_authenticated_unrelated_public_or_authenticated(
+def test_api_documents_descendants_list_authenticated_unrelated_public_or_authenticated(
     reach,
 ):
     """
-    Authenticated users should be able to retrieve the children of a public/authenticated
+    Authenticated users should be able to retrieve the descendants of a public/authenticated
     document to which they are not related.
     """
     user = factories.UserFactory()
@@ -161,14 +202,16 @@ def test_api_documents_children_list_authenticated_unrelated_public_or_authentic
 
     document = factories.DocumentFactory(link_reach=reach)
     child1, child2 = factories.DocumentFactory.create_batch(2, parent=document)
+    grand_child = factories.DocumentFactory(parent=child1)
+
     factories.UserDocumentAccessFactory(document=child1)
 
     response = client.get(
-        f"/api/v1.0/documents/{document.id!s}/children/",
+        f"/api/v1.0/documents/{document.id!s}/descendants/",
     )
     assert response.status_code == 200
     assert response.json() == {
-        "count": 2,
+        "count": 3,
         "next": None,
         "previous": None,
         "results": [
@@ -182,12 +225,30 @@ def test_api_documents_children_list_authenticated_unrelated_public_or_authentic
                 "is_favorite": False,
                 "link_reach": child1.link_reach,
                 "link_role": child1.link_role,
-                "numchild": 0,
+                "numchild": 1,
                 "nb_accesses_ancestors": 1,
                 "nb_accesses_direct": 1,
                 "path": child1.path,
                 "title": child1.title,
                 "updated_at": child1.updated_at.isoformat().replace("+00:00", "Z"),
+                "user_roles": [],
+            },
+            {
+                "abilities": grand_child.get_abilities(user),
+                "created_at": grand_child.created_at.isoformat().replace("+00:00", "Z"),
+                "creator": str(grand_child.creator.id),
+                "depth": 3,
+                "excerpt": grand_child.excerpt,
+                "id": str(grand_child.id),
+                "is_favorite": False,
+                "link_reach": grand_child.link_reach,
+                "link_role": grand_child.link_role,
+                "numchild": 0,
+                "nb_accesses_ancestors": 1,
+                "nb_accesses_direct": 0,
+                "path": grand_child.path,
+                "title": grand_child.title,
+                "updated_at": grand_child.updated_at.isoformat().replace("+00:00", "Z"),
                 "user_roles": [],
             },
             {
@@ -213,11 +274,11 @@ def test_api_documents_children_list_authenticated_unrelated_public_or_authentic
 
 
 @pytest.mark.parametrize("reach", ["public", "authenticated"])
-def test_api_documents_children_list_authenticated_public_or_authenticated_parent(
+def test_api_documents_descendants_list_authenticated_public_or_authenticated_parent(
     reach,
 ):
     """
-    Authenticated users should be allowed to retrieve the children of a document who
+    Authenticated users should be allowed to retrieve the descendants of a document who
     has a public or authenticated ancestor.
     """
     user = factories.UserFactory()
@@ -229,13 +290,15 @@ def test_api_documents_children_list_authenticated_public_or_authenticated_paren
     parent = factories.DocumentFactory(parent=grand_parent, link_reach="restricted")
     document = factories.DocumentFactory(link_reach="restricted", parent=parent)
     child1, child2 = factories.DocumentFactory.create_batch(2, parent=document)
+    grand_child = factories.DocumentFactory(parent=child1)
+
     factories.UserDocumentAccessFactory(document=child1)
 
-    response = client.get(f"/api/v1.0/documents/{document.id!s}/children/")
+    response = client.get(f"/api/v1.0/documents/{document.id!s}/descendants/")
 
     assert response.status_code == 200
     assert response.json() == {
-        "count": 2,
+        "count": 3,
         "next": None,
         "previous": None,
         "results": [
@@ -249,12 +312,30 @@ def test_api_documents_children_list_authenticated_public_or_authenticated_paren
                 "is_favorite": False,
                 "link_reach": child1.link_reach,
                 "link_role": child1.link_role,
-                "numchild": 0,
+                "numchild": 1,
                 "nb_accesses_ancestors": 1,
                 "nb_accesses_direct": 1,
                 "path": child1.path,
                 "title": child1.title,
                 "updated_at": child1.updated_at.isoformat().replace("+00:00", "Z"),
+                "user_roles": [],
+            },
+            {
+                "abilities": grand_child.get_abilities(user),
+                "created_at": grand_child.created_at.isoformat().replace("+00:00", "Z"),
+                "creator": str(grand_child.creator.id),
+                "depth": 5,
+                "excerpt": grand_child.excerpt,
+                "id": str(grand_child.id),
+                "is_favorite": False,
+                "link_reach": grand_child.link_reach,
+                "link_role": grand_child.link_role,
+                "numchild": 0,
+                "nb_accesses_ancestors": 1,
+                "nb_accesses_direct": 0,
+                "path": grand_child.path,
+                "title": grand_child.title,
+                "updated_at": grand_child.updated_at.isoformat().replace("+00:00", "Z"),
                 "user_roles": [],
             },
             {
@@ -279,22 +360,23 @@ def test_api_documents_children_list_authenticated_public_or_authenticated_paren
     }
 
 
-def test_api_documents_children_list_authenticated_unrelated_restricted():
+def test_api_documents_descendants_list_authenticated_unrelated_restricted():
     """
-    Authenticated users should not be allowed to retrieve the children of a document that is
+    Authenticated users should not be allowed to retrieve the descendants of a document that is
     restricted and to which they are not related.
     """
     user = factories.UserFactory(with_owned_document=True)
-
     client = APIClient()
     client.force_login(user)
 
     document = factories.DocumentFactory(link_reach="restricted")
     child1, _child2 = factories.DocumentFactory.create_batch(2, parent=document)
+    _grand_child = factories.DocumentFactory(parent=child1)
+
     factories.UserDocumentAccessFactory(document=child1)
 
     response = client.get(
-        f"/api/v1.0/documents/{document.id!s}/children/",
+        f"/api/v1.0/documents/{document.id!s}/descendants/",
     )
     assert response.status_code == 403
     assert response.json() == {
@@ -302,9 +384,9 @@ def test_api_documents_children_list_authenticated_unrelated_restricted():
     }
 
 
-def test_api_documents_children_list_authenticated_related_direct():
+def test_api_documents_descendants_list_authenticated_related_direct():
     """
-    Authenticated users should be allowed to retrieve the children of a document
+    Authenticated users should be allowed to retrieve the descendants of a document
     to which they are directly related whatever the role.
     """
     user = factories.UserFactory()
@@ -319,12 +401,14 @@ def test_api_documents_children_list_authenticated_related_direct():
     child1, child2 = factories.DocumentFactory.create_batch(2, parent=document)
     factories.UserDocumentAccessFactory(document=child1)
 
+    grand_child = factories.DocumentFactory(parent=child1)
+
     response = client.get(
-        f"/api/v1.0/documents/{document.id!s}/children/",
+        f"/api/v1.0/documents/{document.id!s}/descendants/",
     )
     assert response.status_code == 200
     assert response.json() == {
-        "count": 2,
+        "count": 3,
         "next": None,
         "previous": None,
         "results": [
@@ -338,7 +422,7 @@ def test_api_documents_children_list_authenticated_related_direct():
                 "is_favorite": False,
                 "link_reach": child1.link_reach,
                 "link_role": child1.link_role,
-                "numchild": 0,
+                "numchild": 1,
                 "nb_accesses_ancestors": 3,
                 "nb_accesses_direct": 1,
                 "path": child1.path,
@@ -347,6 +431,24 @@ def test_api_documents_children_list_authenticated_related_direct():
                 "user_roles": [access.role],
             },
             {
+                "abilities": grand_child.get_abilities(user),
+                "created_at": grand_child.created_at.isoformat().replace("+00:00", "Z"),
+                "creator": str(grand_child.creator.id),
+                "depth": 3,
+                "excerpt": grand_child.excerpt,
+                "id": str(grand_child.id),
+                "is_favorite": False,
+                "link_reach": grand_child.link_reach,
+                "link_role": grand_child.link_role,
+                "numchild": 0,
+                "nb_accesses_ancestors": 3,
+                "nb_accesses_direct": 0,
+                "path": grand_child.path,
+                "title": grand_child.title,
+                "updated_at": grand_child.updated_at.isoformat().replace("+00:00", "Z"),
+                "user_roles": [access.role],
+            },
+            {
                 "abilities": child2.get_abilities(user),
                 "created_at": child2.created_at.isoformat().replace("+00:00", "Z"),
                 "creator": str(child2.creator.id),
@@ -368,9 +470,9 @@ def test_api_documents_children_list_authenticated_related_direct():
     }
 
 
-def test_api_documents_children_list_authenticated_related_parent():
+def test_api_documents_descendants_list_authenticated_related_parent():
     """
-    Authenticated users should be allowed to retrieve the children of a document if they
+    Authenticated users should be allowed to retrieve the descendants of a document if they
     are related to one of its ancestors whatever the role.
     """
     user = factories.UserFactory()
@@ -379,22 +481,24 @@ def test_api_documents_children_list_authenticated_related_parent():
     client.force_login(user)
 
     grand_parent = factories.DocumentFactory(link_reach="restricted")
+    grand_parent_access = factories.UserDocumentAccessFactory(
+        document=grand_parent, user=user
+    )
+
     parent = factories.DocumentFactory(parent=grand_parent, link_reach="restricted")
     document = factories.DocumentFactory(parent=parent, link_reach="restricted")
 
     child1, child2 = factories.DocumentFactory.create_batch(2, parent=document)
     factories.UserDocumentAccessFactory(document=child1)
 
-    grand_parent_access = factories.UserDocumentAccessFactory(
-        document=grand_parent, user=user
-    )
+    grand_child = factories.DocumentFactory(parent=child1)
 
     response = client.get(
-        f"/api/v1.0/documents/{document.id!s}/children/",
+        f"/api/v1.0/documents/{document.id!s}/descendants/",
     )
     assert response.status_code == 200
     assert response.json() == {
-        "count": 2,
+        "count": 3,
         "next": None,
         "previous": None,
         "results": [
@@ -408,12 +512,30 @@ def test_api_documents_children_list_authenticated_related_parent():
                 "is_favorite": False,
                 "link_reach": child1.link_reach,
                 "link_role": child1.link_role,
-                "numchild": 0,
+                "numchild": 1,
                 "nb_accesses_ancestors": 2,
                 "nb_accesses_direct": 1,
                 "path": child1.path,
                 "title": child1.title,
                 "updated_at": child1.updated_at.isoformat().replace("+00:00", "Z"),
+                "user_roles": [grand_parent_access.role],
+            },
+            {
+                "abilities": grand_child.get_abilities(user),
+                "created_at": grand_child.created_at.isoformat().replace("+00:00", "Z"),
+                "creator": str(grand_child.creator.id),
+                "depth": 5,
+                "excerpt": grand_child.excerpt,
+                "id": str(grand_child.id),
+                "is_favorite": False,
+                "link_reach": grand_child.link_reach,
+                "link_role": grand_child.link_role,
+                "numchild": 0,
+                "nb_accesses_ancestors": 2,
+                "nb_accesses_direct": 0,
+                "path": grand_child.path,
+                "title": grand_child.title,
+                "updated_at": grand_child.updated_at.isoformat().replace("+00:00", "Z"),
                 "user_roles": [grand_parent_access.role],
             },
             {
@@ -438,24 +560,24 @@ def test_api_documents_children_list_authenticated_related_parent():
     }
 
 
-def test_api_documents_children_list_authenticated_related_child():
+def test_api_documents_descendants_list_authenticated_related_child():
     """
-    Authenticated users should not be allowed to retrieve all the children of a document
+    Authenticated users should not be allowed to retrieve all the descendants of a document
     as a result of being related to one of its children.
     """
     user = factories.UserFactory()
-
     client = APIClient()
     client.force_login(user)
 
     document = factories.DocumentFactory(link_reach="restricted")
     child1, _child2 = factories.DocumentFactory.create_batch(2, parent=document)
+    _grand_child = factories.DocumentFactory(parent=child1)
 
     factories.UserDocumentAccessFactory(document=child1, user=user)
     factories.UserDocumentAccessFactory(document=document)
 
     response = client.get(
-        f"/api/v1.0/documents/{document.id!s}/children/",
+        f"/api/v1.0/documents/{document.id!s}/descendants/",
     )
     assert response.status_code == 403
     assert response.json() == {
@@ -463,15 +585,16 @@ def test_api_documents_children_list_authenticated_related_child():
     }
 
 
-def test_api_documents_children_list_authenticated_related_team_none(mock_user_teams):
+def test_api_documents_descendants_list_authenticated_related_team_none(
+    mock_user_teams,
+):
     """
-    Authenticated users should not be able to retrieve the children of a restricted document
+    Authenticated users should not be able to retrieve the descendants of a restricted document
     related to teams in which the user is not.
     """
     mock_user_teams.return_value = []
 
     user = factories.UserFactory(with_owned_document=True)
-
     client = APIClient()
     client.force_login(user)
 
@@ -480,38 +603,38 @@ def test_api_documents_children_list_authenticated_related_team_none(mock_user_t
 
     factories.TeamDocumentAccessFactory(document=document, team="myteam")
 
-    response = client.get(f"/api/v1.0/documents/{document.id!s}/children/")
+    response = client.get(f"/api/v1.0/documents/{document.id!s}/descendants/")
     assert response.status_code == 403
     assert response.json() == {
         "detail": "You do not have permission to perform this action."
     }
 
 
-def test_api_documents_children_list_authenticated_related_team_members(
+def test_api_documents_descendants_list_authenticated_related_team_members(
     mock_user_teams,
 ):
     """
-    Authenticated users should be allowed to retrieve the children of a document to which they
+    Authenticated users should be allowed to retrieve the descendants of a document to which they
     are related via a team whatever the role.
     """
     mock_user_teams.return_value = ["myteam"]
 
     user = factories.UserFactory()
-
     client = APIClient()
     client.force_login(user)
 
     document = factories.DocumentFactory(link_reach="restricted")
     child1, child2 = factories.DocumentFactory.create_batch(2, parent=document)
+    grand_child = factories.DocumentFactory(parent=child1)
 
     access = factories.TeamDocumentAccessFactory(document=document, team="myteam")
 
-    response = client.get(f"/api/v1.0/documents/{document.id!s}/children/")
+    response = client.get(f"/api/v1.0/documents/{document.id!s}/descendants/")
 
     # pylint: disable=R0801
     assert response.status_code == 200
     assert response.json() == {
-        "count": 2,
+        "count": 3,
         "next": None,
         "previous": None,
         "results": [
@@ -525,12 +648,30 @@ def test_api_documents_children_list_authenticated_related_team_members(
                 "is_favorite": False,
                 "link_reach": child1.link_reach,
                 "link_role": child1.link_role,
-                "numchild": 0,
+                "numchild": 1,
                 "nb_accesses_ancestors": 1,
                 "nb_accesses_direct": 0,
                 "path": child1.path,
                 "title": child1.title,
                 "updated_at": child1.updated_at.isoformat().replace("+00:00", "Z"),
+                "user_roles": [access.role],
+            },
+            {
+                "abilities": grand_child.get_abilities(user),
+                "created_at": grand_child.created_at.isoformat().replace("+00:00", "Z"),
+                "creator": str(grand_child.creator.id),
+                "depth": 3,
+                "excerpt": grand_child.excerpt,
+                "id": str(grand_child.id),
+                "is_favorite": False,
+                "link_reach": grand_child.link_reach,
+                "link_role": grand_child.link_role,
+                "numchild": 0,
+                "nb_accesses_ancestors": 1,
+                "nb_accesses_direct": 0,
+                "path": grand_child.path,
+                "title": grand_child.title,
+                "updated_at": grand_child.updated_at.isoformat().replace("+00:00", "Z"),
                 "user_roles": [access.role],
             },
             {