+++ src/backend/core/tests/documents/test_api_documents_retrieve.py
@@ -34,16 +34,23 @@ def test_api_documents_retrieve_anonymous_public_standalone():
             "children_create": False,
             "children_list": True,
             "collaboration_auth": True,
+            "descendants": True,
             "destroy": False,
             # Anonymous user can't favorite a document even with read access
             "favorite": False,
             "invite_owner": False,
             "link_configuration": False,
+            "link_select_options": {
+                "authenticated": ["reader", "editor"],
+                "public": ["reader", "editor"],
+                "restricted": ["reader", "editor"],
+            },
             "media_auth": True,
             "move": False,
             "partial_update": document.link_role == "editor",
             "restore": False,
             "retrieve": True,
+            "tree": True,
             "update": document.link_role == "editor",
             "versions_destroy": False,
             "versions_list": False,
@@ -57,7 +64,8 @@ def test_api_documents_retrieve_anonymous_public_standalone():
         "is_favorite": False,
         "link_reach": "public",
         "link_role": document.link_role,
-        "nb_accesses": 0,
+        "nb_accesses_ancestors": 0,
+        "nb_accesses_direct": 0,
         "numchild": 0,
         "path": document.path,
         "title": document.title,
@@ -79,6 +87,7 @@ def test_api_documents_retrieve_anonymous_public_parent():
     response = APIClient().get(f"/api/v1.0/documents/{document.id!s}/")
 
     assert response.status_code == 200
+    links = document.get_ancestors().values("link_reach", "link_role")
     assert response.json() == {
         "id": str(document.id),
         "abilities": {
@@ -90,16 +99,19 @@ def test_api_documents_retrieve_anonymous_public_parent():
             "children_create": False,
             "children_list": True,
             "collaboration_auth": True,
+            "descendants": True,
             "destroy": False,
             # Anonymous user can't favorite a document even with read access
             "favorite": False,
             "invite_owner": False,
             "link_configuration": False,
+            "link_select_options": models.LinkReachChoices.get_select_options(links),
             "media_auth": True,
             "move": False,
             "partial_update": grand_parent.link_role == "editor",
             "restore": False,
             "retrieve": True,
+            "tree": True,
             "update": grand_parent.link_role == "editor",
             "versions_destroy": False,
             "versions_list": False,
@@ -113,7 +125,8 @@ def test_api_documents_retrieve_anonymous_public_parent():
         "is_favorite": False,
         "link_reach": document.link_reach,
         "link_role": document.link_role,
-        "nb_accesses": 0,
+        "nb_accesses_ancestors": 0,
+        "nb_accesses_direct": 0,
         "numchild": 0,
         "path": document.path,
         "title": document.title,
@@ -180,15 +193,22 @@ def test_api_documents_retrieve_authenticated_unrelated_public_or_authenticated(
             "children_create": document.link_role == "editor",
             "children_list": True,
             "collaboration_auth": True,
+            "descendants": True,
             "destroy": False,
             "favorite": True,
             "invite_owner": False,
             "link_configuration": False,
+            "link_select_options": {
+                "authenticated": ["reader", "editor"],
+                "public": ["reader", "editor"],
+                "restricted": ["reader", "editor"],
+            },
             "media_auth": True,
             "move": False,
             "partial_update": document.link_role == "editor",
             "restore": False,
             "retrieve": True,
+            "tree": True,
             "update": document.link_role == "editor",
             "versions_destroy": False,
             "versions_list": False,
@@ -202,7 +222,8 @@ def test_api_documents_retrieve_authenticated_unrelated_public_or_authenticated(
         "is_favorite": False,
         "link_reach": reach,
         "link_role": document.link_role,
-        "nb_accesses": 0,
+        "nb_accesses_ancestors": 0,
+        "nb_accesses_direct": 0,
         "numchild": 0,
         "path": document.path,
         "title": document.title,
@@ -232,6 +253,7 @@ def test_api_documents_retrieve_authenticated_public_or_authenticated_parent(rea
     response = client.get(f"/api/v1.0/documents/{document.id!s}/")
 
     assert response.status_code == 200
+    links = document.get_ancestors().values("link_reach", "link_role")
     assert response.json() == {
         "id": str(document.id),
         "abilities": {
@@ -243,15 +265,18 @@ def test_api_documents_retrieve_authenticated_public_or_authenticated_parent(rea
             "children_create": grand_parent.link_role == "editor",
             "children_list": True,
             "collaboration_auth": True,
+            "descendants": True,
             "destroy": False,
             "favorite": True,
             "invite_owner": False,
             "link_configuration": False,
+            "link_select_options": models.LinkReachChoices.get_select_options(links),
             "move": False,
             "media_auth": True,
             "partial_update": grand_parent.link_role == "editor",
             "restore": False,
             "retrieve": True,
+            "tree": True,
             "update": grand_parent.link_role == "editor",
             "versions_destroy": False,
             "versions_list": False,
@@ -265,7 +290,8 @@ def test_api_documents_retrieve_authenticated_public_or_authenticated_parent(rea
         "is_favorite": False,
         "link_reach": document.link_reach,
         "link_role": document.link_role,
-        "nb_accesses": 0,
+        "nb_accesses_ancestors": 0,
+        "nb_accesses_direct": 0,
         "numchild": 0,
         "path": document.path,
         "title": document.title,
@@ -374,7 +400,8 @@ def test_api_documents_retrieve_authenticated_related_direct():
         "is_favorite": False,
         "link_reach": document.link_reach,
         "link_role": document.link_role,
-        "nb_accesses": 2,
+        "nb_accesses_ancestors": 2,
+        "nb_accesses_direct": 2,
         "numchild": 0,
         "path": document.path,
         "title": document.title,
@@ -404,6 +431,7 @@ def test_api_documents_retrieve_authenticated_related_parent():
         f"/api/v1.0/documents/{document.id!s}/",
     )
     assert response.status_code == 200
+    links = document.get_ancestors().values("link_reach", "link_role")
     assert response.json() == {
         "id": str(document.id),
         "abilities": {
@@ -415,15 +443,18 @@ def test_api_documents_retrieve_authenticated_related_parent():
             "children_create": access.role != "reader",
             "children_list": True,
             "collaboration_auth": True,
+            "descendants": True,
             "destroy": access.role == "owner",
             "favorite": True,
             "invite_owner": access.role == "owner",
             "link_configuration": access.role in ["administrator", "owner"],
+            "link_select_options": models.LinkReachChoices.get_select_options(links),
             "media_auth": True,
             "move": access.role in ["administrator", "owner"],
             "partial_update": access.role != "reader",
             "restore": access.role == "owner",
             "retrieve": True,
+            "tree": True,
             "update": access.role != "reader",
             "versions_destroy": access.role in ["administrator", "owner"],
             "versions_list": True,
@@ -437,7 +468,8 @@ def test_api_documents_retrieve_authenticated_related_parent():
         "is_favorite": False,
         "link_reach": "restricted",
         "link_role": document.link_role,
-        "nb_accesses": 2,
+        "nb_accesses_ancestors": 2,
+        "nb_accesses_direct": 0,
         "numchild": 0,
         "path": document.path,
         "title": document.title,
@@ -465,7 +497,8 @@ def test_api_documents_retrieve_authenticated_related_nb_accesses():
         f"/api/v1.0/documents/{document.id!s}/",
     )
     assert response.status_code == 200
-    assert response.json()["nb_accesses"] == 3
+    assert response.json()["nb_accesses_ancestors"] == 3
+    assert response.json()["nb_accesses_direct"] == 1
 
     factories.UserDocumentAccessFactory(document=grand_parent)
 
@@ -473,7 +506,8 @@ def test_api_documents_retrieve_authenticated_related_nb_accesses():
         f"/api/v1.0/documents/{document.id!s}/",
     )
     assert response.status_code == 200
-    assert response.json()["nb_accesses"] == 4
+    assert response.json()["nb_accesses_ancestors"] == 4
+    assert response.json()["nb_accesses_direct"] == 1
 
 
 def test_api_documents_retrieve_authenticated_related_child():
@@ -554,12 +588,10 @@ def test_api_documents_retrieve_authenticated_related_team_members(
     mock_user_teams.return_value = teams
 
     user = factories.UserFactory()
-
     client = APIClient()
     client.force_login(user)
 
     document = factories.DocumentFactory(link_reach="restricted")
-
     factories.TeamDocumentAccessFactory(
         document=document, team="readers", role="reader"
     )
@@ -588,7 +620,8 @@ def test_api_documents_retrieve_authenticated_related_team_members(
         "is_favorite": False,
         "link_reach": "restricted",
         "link_role": document.link_role,
-        "nb_accesses": 5,
+        "nb_accesses_ancestors": 5,
+        "nb_accesses_direct": 5,
         "numchild": 0,
         "path": document.path,
         "title": document.title,
@@ -649,7 +682,8 @@ def test_api_documents_retrieve_authenticated_related_team_administrators(
         "is_favorite": False,
         "link_reach": "restricted",
         "link_role": document.link_role,
-        "nb_accesses": 5,
+        "nb_accesses_ancestors": 5,
+        "nb_accesses_direct": 5,
         "numchild": 0,
         "path": document.path,
         "title": document.title,
@@ -710,7 +744,8 @@ def test_api_documents_retrieve_authenticated_related_team_owners(
         "is_favorite": False,
         "link_reach": "restricted",
         "link_role": document.link_role,
-        "nb_accesses": 5,
+        "nb_accesses_ancestors": 5,
+        "nb_accesses_direct": 5,
         "numchild": 0,
         "path": document.path,
         "title": document.title,
@@ -719,7 +754,7 @@ def test_api_documents_retrieve_authenticated_related_team_owners(
     }
 
 
-def test_api_documents_retrieve_user_roles(django_assert_num_queries):
+def test_api_documents_retrieve_user_roles(django_assert_max_num_queries):
     """
     Roles should be annotated on querysets taking into account all documents ancestors.
     """
@@ -744,7 +779,7 @@ def test_api_documents_retrieve_user_roles(django_assert_num_queries):
     )
     expected_roles = {access.role for access in accesses}
 
-    with django_assert_num_queries(10):
+    with django_assert_max_num_queries(12):
         response = client.get(f"/api/v1.0/documents/{document.id!s}/")
 
     assert response.status_code == 200
@@ -761,7 +796,7 @@ def test_api_documents_retrieve_numqueries_with_link_trace(django_assert_num_que
 
     document = factories.DocumentFactory(users=[user], link_traces=[user])
 
-    with django_assert_num_queries(4):
+    with django_assert_num_queries(5):
         response = client.get(f"/api/v1.0/documents/{document.id!s}/")
 
     with django_assert_num_queries(3):